Introduces specialised generator for vote structs
